Use Reddit for comments with missing parents
Pushshift occasionally fails to return comments which are the parents
of comments that it does return. Look these parents up in the Reddit
API and use its results instead, otherwise entire comment trees may be
hidden. They could still be hidden if two consecutive parents are
missing from Pushshift, but hopefully that isn't a common occurance.