Add '?max_comments=n' query parameter
This increases the initial max-comments-to-download above the user's
preference without changing their saved setting. This can be useful when
linking to a specific comment which is new enough that the default max-
comments setting (of 800) is too low to download the linked comment
tree, causing most users to see a confusing 'No comments found' message.